Fix Unicorn + Install Store

+++ imentore-admin/app/models/admin_imentore/store.rb
@@ -4,7 +4,7 @@
 module AdminImentore
   class Store
 
-    def self.fix_utf8(text)
+    def self.fix_utf8(text)      
       text.encode('cp1252',
                   :fallback => {
                     "\u0081" => "\x81".force_encoding("cp1252"),
@@ -13,7 +13,7 @@ module AdminImentore
                     "\u0090" => "\x90".force_encoding("cp1252"),
                     "\u009D" => "\x9D".force_encoding("cp1252")
                   })
-          ..encode("cp1252").force_encoding("utf-8")
+          .force_encoding("utf-8")
     end
 
     def self.install_store(store)
@@ -43,10 +43,10 @@ module AdminImentore
       new_store.products.destroy_all
       store.products.not_deleted.sellable.each do |product|
         new_product = new_store.products.new
-        new_product.name = fix_utf8(product.name)
-        new_product.description = fix_utf8(product.description)
+        new_product.name = product.name
+        new_product.description = product.description
         new_product.active = product.sellable
-        new_product.handle = ActiveSupport::Inflector.transliterate(fix_utf8(product.name)).to_underscore
+        new_product.handle = ActiveSupport::Inflector.transliterate(product.name).to_underscore
         new_product.created_at = product.created_at
         new_product.save
         unless new_product.save
@@ -61,7 +61,7 @@ module AdminImentore
         end
 
         product.categories.each do |category|
-          c = new_store.categories.find_by_name(fix_utf8(category.title))
+          c = new_store.categories.find_by_name(category.title)
           unless c.nil?
             c = c.categories_products.new
             c.product_id = new_product.id
@@ -92,17 +92,40 @@ module AdminImentore
               next
             end
           end
-        # end
+        
 
       end
     end      
 
-    def self.install_store_admin(store, new_store)
+    def self.install_customers_employees(store, new_store)
       return if store.nil?
+      store.customers.each do |customer|
+        unless customer.encrypted_password.nil?
+          new_customer = new_store.customers.new 
+          new_customer.name = customer.name
+          new_customer.brand = customer.name if customer.type == "Person"
+          new_customer.irs_id = customer.irs_id 
+          new_customer.national_id = customer.national_id
+          new_customer.birthdate = customer.birthdate
+          new_customer.gender = customer.gender
+          new_customer.person_type = customer.type == 'Person' ? 'person' : 'company'
+          new_customer.store_id = new_store.id
+          new_customer.active = customer.active
+          new_customer.created_at = customer.created_at
+          new_customer.save
+          user = new_customer.build_user(email: customer.email, encrypted_password: customer.encrypted_password, 
+                                   sign_in_count: customer.sign_in_count, current_sign_in_at: customer.current_sign_in_at,
+                                   last_sign_in_ip: customer.last_sign_in_ip, confirmed_at: customer.confirmed_at, 
+                                   created_at: customer.created_at, store_id: new_store.id, password_required: false, 
+                                   password_salt: customer.password_salt)
+          new_customer.destroy unless user.save(validate: false) 
+        end
+      end
+
       employee = store.users.find_by_role('admin')
       new_employee = new_store.employees.new 
-      new_employee.name = fix_utf8(employee.name)
-      new_employee.brand = fix_utf8(employee.name) if employee.type == "Person"
+      new_employee.name = employee.name
+      new_employee.brand = employee.name if employee.type == "Person"
       new_employee.irs_id = employee.irs_id 
       new_employee.national_id = employee.national_id
       new_employee.birthdate = employee.birthdate
@@ -120,14 +143,12 @@ module AdminImentore
                                last_sign_in_ip: employee.last_sign_in_ip, confirmed_at: employee.confirmed_at, 
                                created_at: employee.created_at, store_id: new_store.id, password_required: false,
                                password_salt: employee.password_salt)
-      unless user.save
-        new_employee.destroy
-      end
+      new_employee.destroy unless user.save(validate: false)
     end
 
     def self.category_create(new_category, category, store)
       category.children.each do |children|
-        category_1 = new_category.children.new(name: fix_utf8(children.title), handle: ActiveSupport::Inflector.transliterate(fix_utf8(children.title)).to_underscore, store_id: store.id)
+        category_1 = new_category.children.new(name: children.title, handle: ActiveSupport::Inflector.transliterate(children.title).to_underscore, store_id: store.id)
         unless category_1.save
           category_1.handle = category_1.handle + "_#{rand(1000)}"
           unless category_1.save
@@ -142,9 +163,9 @@ module AdminImentore
 
     def self.install(store)
       new_store = Imentore::Store.new
-      new_store.name = fix_utf8(store.name) 
-      new_store.brand = fix_utf8(store.brand) unless store.brand.blank?
-      new_store.brand ||= fix_utf8(store.name) 
+      new_store.name = store.name 
+      new_store.brand = store.brand unless store.brand.blank?
+      new_store.brand ||= store.name 
       new_store.url = store.url
       new_store.irs_id = store.irs_id
       new_store.plan_id = store.plan_id
@@ -165,7 +186,7 @@ module AdminImentore
       end
 
       store.categories.roots.each do |category|
-        new_category = new_store.categories.new(name: fix_utf8(category.title), handle: ActiveSupport::Inflector.transliterate(fix_utf8(category.title)).to_underscore)
+        new_category = new_store.categories.new(name: category.title, handle: ActiveSupport::Inflector.transliterate(category.title).to_underscore)
         unless new_category.save
           new_category.handle = new_category.handle + "_#{rand(1000)}"
           new_category.store_id = new_store.id
@@ -195,45 +216,19 @@ module AdminImentore
       store.notices.each do |page|
         new_page = new_store.notices.new
         new_page.active = page.active
-        new_page.name = fix_utf8(page.title)
-        new_page.handle = ActiveSupport::Inflector.transliterate(fix_utf8(page.title)).to_underscore
-        new_page.body = fix_utf8(page.description)
+        new_page.name = page.title
+        new_page.handle = ActiveSupport::Inflector.transliterate(page.title).to_underscore
+        new_page.body = page.description
         new_page.created_at = page.created_at
       end
 
-
-      store.customers.each do |customer|
-        unless customer.encrypted_password.nil?
-          new_customer = new_store.customers.new 
-          new_customer.name = fix_utf8(customer.name)
-          new_customer.brand = fix_utf8(customer.name) if customer.type == "Person"
-          new_customer.irs_id = customer.irs_id 
-          new_customer.national_id = customer.national_id
-          new_customer.birthdate = customer.birthdate
-          new_customer.gender = customer.gender
-          new_customer.person_type = customer.type == 'Person' ? 'person' : 'company'
-          new_customer.store_id = new_store.id
-          new_customer.active = customer.active
-          new_customer.created_at = customer.created_at
-          new_customer.save
-          user = new_customer.build_user(email: customer.email, encrypted_password: customer.encrypted_password, 
-                                   sign_in_count: customer.sign_in_count, current_sign_in_at: customer.current_sign_in_at,
-                                   last_sign_in_ip: customer.last_sign_in_ip, confirmed_at: customer.confirmed_at, 
-                                   created_at: customer.created_at, store_id: new_store.id, password_required: false, 
-                                   password_salt: customer.password_salt)
-          unless user.save
-            new_customer.destroy
-          end
-        end
-      end
-
-      install_store_admin(store, new_store)
+      install_customers_employees(store, new_store)
 
       store.employees.each do |employee|
         unless employee.encrypted_password.nil?
           new_employee = new_store.employees.new 
-          new_employee.name = fix_utf8(employee.name)
-          new_employee.brand = fix_utf8(employee.name) if employee.type == "Person"
+          new_employee.name = employee.name
+          new_employee.brand = employee.name if employee.type == "Person"
           new_employee.irs_id = employee.irs_id 
           new_employee.national_id = employee.national_id
           new_employee.birthdate = employee.birthdate
@@ -284,10 +279,10 @@ module AdminImentore
       store.products.not_deleted.each do |product|
         new_product = Imentore::Product.new
         new_product.store_id = new_store.id
-        new_product.name = fix_utf8(product.name)
-        new_product.description = fix_utf8(product.description)
+        new_product.name = product.name
+        new_product.description = product.description
         new_product.active = product.sellable
-        new_product.handle = ActiveSupport::Inflector.transliterate(fix_utf8(product.name)).to_underscore
+        new_product.handle = ActiveSupport::Inflector.transliterate(product.name).to_underscore
         new_product.store_id = new_store.id
         new_product.save
         unless new_product.save
@@ -300,7 +295,7 @@ module AdminImentore
         end
 
         product.categories.each do |category|
-          c = new_store.categories.find_by_name(fix_utf8(category.title))
+          c = new_store.categories.find_by_name(category.title)
           unless c.nil?
             c = c.categories_products.new
             c.product_id = new_product.id
@@ -321,9 +316,6 @@ module AdminImentore
           end
           new_variant.options.create(option_type: default_option, value: ActiveSupport::Inflector.transliterate(variant.name).to_underscore)
         end
-
-
-
 
         unless Rails.env == "development"
           product.images.each do |image|
